Move ki hair to character-fx instead of metaverse modules
