+++ metaverse_modules/ki/index.js
@@ -366,89 +366,8 @@ export default e => {
   let nextCapsuleParticleTime = 0;
   let nextAuraParticleTime = 0;
 
-  const _makeKiHairMaterial = () => {
-    let wVertex = THREE.ShaderLib["standard"].vertexShader;
-    let wFragment = THREE.ShaderLib["standard"].fragmentShader;
-    // let wUniforms = THREE.UniformsUtils.clone(THREE.ShaderLib["standard"].uniforms);
-    let wUniforms = {
-      iTime: {
-        value: 0,
-        needsUpdate: false,
-      },
-      uHeadCenter: {
-        value: new THREE.Vector3(0, 0, 0),
-        needsUpdate: false,
-      },
-      color1: {
-        value: new THREE.Color(),
-        needsUpdate: false,
-      },
-      color2: {
-        value: new THREE.Color(),
-        needsUpdate: false,
-      },
-    };
-    wVertex = wVertex.replace(`#include <clipping_planes_pars_vertex>`, `\
-      #include <clipping_planes_pars_vertex>
-      varying vec2 vUv;
-      varying vec3 vWorldPosition;
-    `);
-    wVertex = wVertex.replace(`}`, `\
-        vUv = uv;
-      }
-    `);
-    wVertex = wVertex.replace(`#include <skinning_vertex>`, `\
-      #include <skinning_vertex>
-      vWorldPosition = transformed;
-    `);
-
-    wFragment = wFragment.replace(`#include <clipping_planes_pars_fragment>`, `\
-      #include <clipping_planes_pars_fragment>
-      uniform float iTime;
-      uniform vec3 uHeadCenter;
-      uniform vec3 color1;
-      uniform vec3 color2;
-      varying vec2 vUv;
-      varying vec3 vWorldPosition;
-
-      // #define PI 3.1415926535897932384626433832795
-    `);
-    wFragment = wFragment.replace(`}`, `\
-        // float f = dot(vNormal, normalize(vec3(1.)));
-        // vec3 colorX = mix(color1, color2, f);
-        // vec3 colorY = mix(color1, color2, vUv.y);
-        // gl_FragColor.rgb = colorX * colorY;
-        // gl_FragColor.rgb = colorX;
-        gl_FragColor.rb = vUv;
-
-        float distanceToCenter = length(vWorldPosition - uHeadCenter);
-        float glowRadius = iTime * 0.15;
-        float distanceToGlowRadius = abs(distanceToCenter - glowRadius);
-        distanceToGlowRadius = mod(distanceToGlowRadius, 0.1);
-        gl_FragColor.rgb *= 0.3;
-        if (distanceToGlowRadius < 0.01) {
-          gl_FragColor.rgb *= 2.;
-        }
-        float glowFactor = 1. + sin(iTime * PI * 2. * 3.) * 0.5;
-        gl_FragColor.rgb *= 0.5 + glowFactor * 0.5;
-        gl_FragColor.a = 1.;
-      }
-    `);
-
-    const material = new THREE.ShaderMaterial({
-      uniforms: wUniforms,
-      vertexShader: wVertex,
-      fragmentShader: wFragment,
-      extensions: {
-        derivatives: true,
-      },
-      // side: THREE.DoubleSide,
-    });
-    return material;
-  };
-
   // const timeOffset = Math.random() * 10;
-  let hairMeshes = null;
+  // let hairMeshes = null;
   useFrame(({timestamp, timeDiff}) => {
     const localPlayer = useLocalPlayer();
     const now = timestamp;
@@ -552,92 +471,6 @@ export default e => {
       startTimesAttribute.needsUpdate = true;
       aura.count = auraParticles.length;
     };
-
-    if (localPlayer.avatar) {
-      // window.avatar = localPlayer.avatar;
-      const {object, vrmExtension} = localPlayer.avatar;
-      
-      /* const {parser} = object;
-      const {json} = parser;
-      const {nodes} = json;
-
-      const {secondaryAnimation} = vrmExtension;
-      const {boneGroups} = secondaryAnimation;
-      const {bones} = boneGroups; */
-
-      // const hairBoneBooleans = bones.map(bone => /hair/i.test(bone.name));
-      if (!object.kiDecorated) {
-        hairMeshes = [];
-        object.scene.traverse(o => {
-          // console.log(o.name, o.isMesh);
-          if (o.isSkinnedMesh) {
-            const {geometry, skeleton} = o;
-            const {attributes, index: indexAttribute} = geometry;
-            const indices = indexAttribute.array;
-            const {skinIndex, skinWeight} = attributes;
-            const skinIndices = skinIndex.array;
-            const skinWeights = skinWeight.array;
-            const {itemSize} = skinIndex;
-            let done = false;
-            for (let i = 0; i < indices.length; i++) {
-              const index = indices[i];
-              for (let j = 0; j < itemSize; j++) {
-                const skinIndex = skinIndices[index * itemSize + j];
-                const skinWeight = skinWeights[index * itemSize + j];
-                if (skinWeight !== 0 && /hair/i.test(skeleton.bones[skinIndex].name)) {
-                  // console.log('got', o.name, {attributes, skinIndices, skinWeights});
-                  hairMeshes.push(o);
-                  done = true;
-                  break;
-                }
-              }
-              if (done) {
-                break;
-              }
-            }
-          }
-        });
-        for (const hairMesh of hairMeshes) {
-          hairMesh.kiOriginalMaterial = hairMesh.material;
-          const kiHairMaterial = _makeKiHairMaterial(hairMesh.material);
-          hairMesh.material = kiHairMaterial;
-          
-          kiHairMaterial.uniforms.color1.value.setHex(0xfdeb44);
-          kiHairMaterial.uniforms.color1.needsUpdate = true;
-          kiHairMaterial.uniforms.color2.value.setHex(0xf6b01d);
-          kiHairMaterial.uniforms.color2.needsUpdate = true;
-
-          for (const k in kiHairMaterial.uniforms) {
-            hairMesh.material.uniforms[k] = kiHairMaterial.uniforms[k];
-          }
-          hairMesh.material.defines = kiHairMaterial.defines;
-          hairMesh.material.vertexShader = kiHairMaterial.vertexShader;
-          hairMesh.material.fragmentShader = kiHairMaterial.fragmentShader;
-        }
-        object.kiDecorated = true;
-      }
-      for (const hairMesh of hairMeshes) {
-        hairMesh.material.uniforms.iTime.value = timestamp/1000;
-        hairMesh.material.uniforms.iTime.needsUpdate = true;
-
-        hairMesh.material.uniforms.uHeadCenter.value.setFromMatrixPosition(localPlayer.avatar.modelBoneOutputs.Head.matrixWorld);
-        hairMesh.material.uniforms.uHeadCenter.needsUpdate = true;
-      }
-
-      if (localPlayer.avatar.springBoneManager) {
-        localPlayer.avatar.springBoneManager.springBoneGroupList[0].forEach(o => {
-          o.gravityDir.setFromMatrixPosition(o.bone.matrixWorld)
-            .sub(localVector.setFromMatrixPosition(localPlayer.avatar.modelBoneOutputs.Head.matrixWorld))
-            .normalize()
-            .lerp(localVector.set(0, 1, 0), 0.9)
-            // .applyQuaternion(localQuaternion.setFromRotationMatrix(o.bone.matrixWorld));
-          const octave = Math.sin(timeS * Math.PI * 2 * 4) // +
-            // Math.sin(timeS * Math.PI * 2 * 4) +
-            // Math.sin(timeS * Math.PI * 2 * 8);
-          o.gravityPower = 0.4 + (1 + octave)*0.5 * 0.5;
-        });
-      }
-    }
 
     if (now >= nextGroundWindParticleTime) {
       const position = new THREE.Vector3(Math.random() * 2 - 1, Math.random() * 2 - 1, Math.random() * 2 - 1)
